+++ EddiInaraService/InaraService.cs
@@ -84,7 +84,7 @@ namespace EddiInaraService
         }
 
         // If you need to do some testing on Inara's API, please set the `inBeta` boolean header property to true.
-        public List<InaraResponse> SendEventBatch(ref List<InaraAPIEvent> events, bool inBeta)
+        public List<InaraResponse> SendEventBatch(ref List<InaraAPIEvent> events, bool inBeta = false)
         {
             List<InaraResponse> inaraResponses = new List<InaraResponse>();
 
@@ -162,7 +162,7 @@ namespace EddiInaraService
             }
         }
 
-        public async void SendQueuedAPIEventsAsync(bool inBeta)
+        public async void SendQueuedAPIEventsAsync(bool inBeta = false)
         {
             List<InaraAPIEvent> queue = new List<InaraAPIEvent>();
             while (Instance.queuedAPIEvents.TryDequeue(out InaraAPIEvent pendingEvent))