+++ EDDI/ChangeLog.md
@@ -26,6 +26,8 @@ Full details of the variables available for each noted event, and VoiceAttack in
     * Updated the `System report` script to enhance the description of powerplay status.
     * Updated the `Engineer progressed` event to stay silent by default when written at startup (with empty values, signaling that engineer data has been loaded).
     * Updated the `Bodies mapped` script to correct a typo.
+  * Speech service
+    * Added new function `InaraDetails` for looking up commander details on https://inara.cz. 
   * VoiceAttack responder
     * Added new plugin function `inara`, allowing commanders to look up the Inara profiles of other commanders in their browsers.